move rho calculations to base class
